try to fix Inexact Error in win32

+++ src/Messages.jl
@@ -75,7 +75,7 @@ using ..IOExtras
 using ..Parsers
 import ..Parsers: headerscomplete, reset!
 
-const unknown_length = typemax(Int64)
+const unknown_length = typemax(Int)
 
 
 abstract type Message end
